+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 ### Changed
 * `multidispatch` supports variable keyword arguments
+* `multidispatch` base implementation always matches
 
 ## [1.11.2](https://pypi.org/project/multimethod/1.11.2/) - 2024-02-27
 ### Fixed